# ai_tnhn/ai-api-tnhn/docs/swagger.yaml
basePath: /api
definitions:
...
  web.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: error message
        type: string
      status:
        example: error
        type: string
    type: object
  web.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      status:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Máy chủ API cho Hệ thống Giám sát Ngập lụt Hà Nội.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Hệ thống Giám sát Ngập lụt Hà Nội
  version: "1.0"
paths:
...
  /water/lake/{id}/history:
    get:
      description: Truy xuất dữ liệu mực nước hồ lịch sử của một trạm cụ thể
      parameters:
      - description: ID trạm
        in: path
        name: id
        required: true
        type: integer
      - default: 100
        description: Số bản ghi tối đa
        in: query
        name: limit
        type: integer
      - description: Lọc theo ngày (YYYY-MM-DD)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Lấy lịch sử dữ liệu trạm hồ
      tags:
      - Thủy văn
  /water/rain/{id}/history:
    get:
      description: Truy xuất dữ liệu mưa lịch sử của một trạm cụ thể
      parameters:
      - description: ID trạm
        in: path
        name: id
        required: true
        type: integer
      - default: 100
        description: Số bản ghi tối đa
        in: query
        name: limit
        type: integer
      - description: Lọc theo ngày (YYYY-MM-DD)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Lấy lịch sử dữ liệu trạm mưa
      tags:
      - Thủy văn
  /water/rain/by-date:
    get:
      description: Truy xuất dữ liệu mưa của tất cả các trạm trong một ngày cụ thể
      parameters:
      - description: Ngày lọc (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Lấy dữ liệu mưa tất cả trạm theo ngày
      tags:
      - Thủy văn
  /water/river/{id}/history:
    get:
      description: Truy xuất dữ liệu mực nước sông lịch sử của một trạm cụ thể
      parameters:
      - description: ID trạm
        in: path
        name: id
        required: true
        type: integer
      - default: 100
        description: Số bản ghi tối đa
        in: query
        name: limit
        type: integer
      - description: Lọc theo ngày (YYYY-MM-DD)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Lấy lịch sử dữ liệu trạm sông
      tags:
      - Thủy văn
  /weather/rain:
    get:
      description: Truy xuất dữ liệu lượng mưa hiện tại của tất cả các trạm quan trắc
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy tổng hợp lượng mưa hiện tại
      tags:
      - Thời tiết
  /weather/rain/comparison:
    get:
      description: Truy xuất báo cáo so sánh dữ liệu lượng mưa giữa hai năm cụ thể
      parameters:
      - description: Năm thứ nhất
        in: query
        name: year1
        required: true
        type: integer
      - description: Năm thứ hai
        in: query
        name: year2
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Báo cáo so sánh lượng mưa giữa hai năm
      tags:
      - Thời tiết
  /weather/rain/history:
    get:
      description: Truy xuất dữ liệu lượng mưa lịch sử của tất cả các trạm
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy dữ liệu lịch sử lượng mưa
      tags:
      - Thời tiết
  /weather/water:
    get:
      description: Truy xuất dữ liệu mực nước hiện tại của tất cả các trạm quan trắc
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy tổng hợp mực nước hiện tại
      tags:
      - Thời tiết
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"